micropython/utarfile: @jimmo's reorganization and optimizations.
